combine unit tests into node/internal/fs/streams_test.ts
