+++ node/internal/fs/testdata/hello.txt
@@ -1,0 +1,1 @@
+hello world
